Fixed bug where activation payments could be negative.
